+++ open-sse/services/tokenRefresh/providers/copilot.ts
@@ -1,0 +1,44 @@
+// @ts-nocheck
+// Extracted from open-sse/services/tokenRefresh.ts — see ../shared.ts for
+// provenance notes (ported idea from KooshaPari's PR #7338, redone on tip).
+import { getGitHubCopilotRefreshHeaders } from "../../../config/providerHeaderProfiles.ts";
+import { runWithProxyContext } from "../../../utils/proxyFetch.ts";
+
+/**
+ * Refresh GitHub Copilot token using GitHub access token
+ */
+export async function refreshCopilotToken(githubAccessToken, log, proxyConfig: unknown = null) {
+  try {
+    const response = await runWithProxyContext(proxyConfig, () =>
+      fetch("https://api.github.com/copilot_internal/v2/token", {
+        headers: getGitHubCopilotRefreshHeaders(`token ${githubAccessToken}`),
+      })
+    );
+
+    if (!response.ok) {
+      const errorText = await response.text();
+      log?.error?.("TOKEN_REFRESH", "Failed to refresh Copilot token", {
+        status: response.status,
+        error: errorText,
+      });
+      return null;
+    }
+
+    const data = await response.json();
+
+    log?.info?.("TOKEN_REFRESH", "Successfully refreshed Copilot token", {
+      hasToken: !!data.token,
+      expiresAt: data.expires_at,
+    });
+
+    return {
+      token: data.token,
+      expiresAt: data.expires_at,
+    };
+  } catch (error) {
+    log?.error?.("TOKEN_REFRESH", "Error refreshing Copilot token", {
+      error: error.message,
+    });
+    return null;
+  }
+}
