+++ tests/unit/oauth-providers-error-handling.test.ts
@@ -29,7 +29,10 @@ test("P0: gitlab-duo is registered in providerRegistry", async () => {
 });
 
 test("P0: refreshGitLabDuoToken exists and handles invalid_grant as unrecoverable", async () => {
-  const src = await read("open-sse/services/tokenRefresh.ts");
+  // refreshGitLabDuoToken lives in its own co-located provider module since the
+  // tokenRefresh.ts provider-extraction (originally proposed in #7338, redone
+  // on tip) — tokenRefresh.ts now only re-exports it.
+  const src = await read("open-sse/services/tokenRefresh/providers/gitlabDuo.ts");
   assert.match(
     src,
     /export\s+async\s+function\s+refreshGitLabDuoToken\(/,
@@ -59,7 +62,9 @@ test("P0: gitlab-duo is in supportsTokenRefresh explicit set", async () => {
 // ─── P1: Kimi Coding stable device_id ────────────────────────────────────────
 
 test("P1: refreshKimiCodingToken accepts providerSpecificData parameter", async () => {
-  const src = await read("open-sse/services/tokenRefresh.ts");
+  // refreshKimiCodingToken lives in its own co-located provider module since
+  // the tokenRefresh.ts provider-extraction (#7338, redone on tip).
+  const src = await read("open-sse/services/tokenRefresh/providers/kimiCoding.ts");
   assert.match(
     src,
     /export\s+async\s+function\s+refreshKimiCodingToken\([^)]*providerSpecificData/,
@@ -68,7 +73,7 @@ test("P1: refreshKimiCodingToken accepts providerSpecificData parameter", async 
 });
 
 test("P1: refreshKimiCodingToken does NOT use ephemeral Date.now() device ID", async () => {
-  const src = await read("open-sse/services/tokenRefresh.ts");
+  const src = await read("open-sse/services/tokenRefresh/providers/kimiCoding.ts");
   // Extract function body — match from declaration to next top-level export function
   const fnMatch = src.match(/export\s+async\s+function\s+refreshKimiCodingToken\([\s\S]+?\n\}/);
   assert.ok(fnMatch, "refreshKimiCodingToken function body not found");
@@ -80,7 +85,7 @@ test("P1: refreshKimiCodingToken does NOT use ephemeral Date.now() device ID", a
 });
 
 test("P1: refreshKimiCodingToken handles invalid_grant as unrecoverable", async () => {
-  const src = await read("open-sse/services/tokenRefresh.ts");
+  const src = await read("open-sse/services/tokenRefresh/providers/kimiCoding.ts");
   const fnMatch = src.match(/export\s+async\s+function\s+refreshKimiCodingToken\([\s\S]+?\n\}/);
   assert.ok(fnMatch, "refreshKimiCodingToken function body not found");
   assert.match(fnMatch[0], /invalid_grant/, "must detect invalid_grant");
@@ -159,7 +164,9 @@ test("P1: GitHub Copilot sub-token guard normalizes conn.provider case", async (
 // ─── P2: Google invalid_grant ─────────────────────────────────────────────────
 
 test("P2: refreshGoogleToken parses invalid_grant as unrecoverable", async () => {
-  const src = await read("open-sse/services/tokenRefresh.ts");
+  // refreshGoogleToken lives in its own co-located provider module since the
+  // tokenRefresh.ts provider-extraction (#7338, redone on tip).
+  const src = await read("open-sse/services/tokenRefresh/providers/google.ts");
   const fnMatch = src.match(/export\s+async\s+function\s+refreshGoogleToken\([\s\S]+?\n\}/);
   assert.ok(fnMatch, "refreshGoogleToken function body not found");
   assert.match(fnMatch[0], /invalid_grant/, "must detect invalid_grant");
@@ -169,7 +176,9 @@ test("P2: refreshGoogleToken parses invalid_grant as unrecoverable", async () =>
 // ─── P2: Qwen invalid_grant ───────────────────────────────────────────────────
 
 test("P2: refreshQwenToken handles invalid_grant in addition to invalid_request", async () => {
-  const src = await read("open-sse/services/tokenRefresh.ts");
+  // refreshQwenToken lives in its own co-located provider module since the
+  // tokenRefresh.ts provider-extraction (#7338, redone on tip).
+  const src = await read("open-sse/services/tokenRefresh/providers/qwen.ts");
   const fnMatch = src.match(/export\s+async\s+function\s+refreshQwenToken\([\s\S]+?\n\}/);
   assert.ok(fnMatch, "refreshQwenToken function body not found");
   assert.match(fnMatch[0], /invalid_grant/, "must detect invalid_grant");
@@ -179,7 +188,9 @@ test("P2: refreshQwenToken handles invalid_grant in addition to invalid_request"
 // ─── P2: Kiro AWS InvalidGrantException ──────────────────────────────────────
 
 test("P2: refreshKiroToken parses AWS InvalidGrantException", async () => {
-  const src = await read("open-sse/services/tokenRefresh.ts");
+  // refreshKiroToken lives in its own co-located provider module since the
+  // tokenRefresh.ts provider-extraction (#7338, redone on tip).
+  const src = await read("open-sse/services/tokenRefresh/providers/kiro.ts");
   const fnMatch = src.match(/export\s+async\s+function\s+refreshKiroToken\([\s\S]+?\n\}/);
   assert.ok(fnMatch, "refreshKiroToken function body not found");
   assert.match(
@@ -191,7 +202,7 @@ test("P2: refreshKiroToken parses AWS InvalidGrantException", async () => {
 });
 
 test("P2: refreshKiroToken handles AWS errors on both AWS OIDC and social auth paths", async () => {
-  const src = await read("open-sse/services/tokenRefresh.ts");
+  const src = await read("open-sse/services/tokenRefresh/providers/kiro.ts");
   const fnMatch = src.match(/export\s+async\s+function\s+refreshKiroToken\([\s\S]+?\n\}/);
   assert.ok(fnMatch, "refreshKiroToken function body not found");
   // Count occurrences of InvalidGrantException — should appear in both paths
@@ -205,7 +216,9 @@ test("P2: refreshKiroToken handles AWS errors on both AWS OIDC and social auth p
 // ─── P3: Claude error shape normalization ─────────────────────────────────────
 
 test("P3: refreshClaudeOAuthToken normalizes invalid_grant to unrecoverable_refresh_error sentinel", async () => {
-  const src = await read("open-sse/services/tokenRefresh.ts");
+  // refreshClaudeOAuthToken lives in its own co-located provider module since
+  // the tokenRefresh.ts provider-extraction (#7338, redone on tip).
+  const src = await read("open-sse/services/tokenRefresh/providers/claudeOAuth.ts");
   const fnMatch = src.match(/export\s+async\s+function\s+refreshClaudeOAuthToken\([\s\S]+?\n\}/);
   assert.ok(fnMatch, "refreshClaudeOAuthToken function body not found");
   assert.match(
@@ -224,7 +237,9 @@ test("P3: refreshClaudeOAuthToken normalizes invalid_grant to unrecoverable_refr
 // ─── P3: Windsurf Firebase errors ────────────────────────────────────────────
 
 test("P3: refreshWindsurfToken parses Firebase USER_DISABLED/TOKEN_EXPIRED errors", async () => {
-  const src = await read("open-sse/services/tokenRefresh.ts");
+  // refreshWindsurfToken lives in its own co-located provider module since the
+  // tokenRefresh.ts provider-extraction (#7338, redone on tip).
+  const src = await read("open-sse/services/tokenRefresh/providers/windsurf.ts");
   const fnMatch = src.match(/export\s+async\s+function\s+refreshWindsurfToken\([\s\S]+?\n\}/);
   assert.ok(fnMatch, "refreshWindsurfToken function body not found");
   assert.match(
